clean code and var names
